+++ app/logic/Mail/SQL/Account/SQLDatabase.ts
@@ -35,11 +35,16 @@ export async function makeTestDatabase(): Promise<Database> {
   return accountsDatabase;
 }
 
-async function deleteDatabase(): Promise<void> {
+export async function deleteDatabase(): Promise<void> {
   let tables = await accountsDatabase.all(sql`SELECT name FROM sqlite_schema WHERE type='table'`) as any[];
   for (let row of tables) {
     let table = row.name;
+    if (table?.startsWith("sqlite_")) {
+      continue;
+    }
     await accountsDatabase.execute(sql`DROP TABLE IF EXISTS ${table};`);
   }
   await accountsDatabase.pragma('user_version = 0');
+  (accountsDatabase as any).close(); // any, because function exists, but not in TypeScript definitions
+  accountsDatabase = null;
 }
